+++ packages/fighting-design/_interface/event.d.ts
@@ -4,4 +4,14 @@
  * @see MouseEvent https://developer.mozilla.org/zh-CN/docs/Web/API/MouseEvent/MouseEvent
  * @param evt 事件对象
  */
-export type HandelClick = (evt: MouseEvent) => void
+export type HandleMouse = (evt: MouseEvent) => void
+
+/**
+ * 普通事件参数类型接口
+ * 
+ * @see Event https://developer.mozilla.org/zh-CN/docs/Web/API/Event
+ * @param evt 事件对象
+ */
+export type HandleEvent = (evt: Event) => void
+
+export type HandleChange = (target: boolean) => void