+++ packages/fighting-design/close-btn/src/props.ts
@@ -1,5 +1,5 @@
 import type { ExtractPropTypes, PropType } from 'vue'
-import type { HandleEvent, FightingIcon } from '../../_interface'
+import type { HandleMouse, FightingIcon } from '../../_interface'
 
 export const Props = {
   /**
@@ -41,7 +41,7 @@ export const Props = {
    * 点击之后执行的回调
    */
   onClick: {
-    type: Function as PropType<HandleEvent>,
+    type: Function as PropType<HandleMouse>,
     default: (): null => null
   }
 } as const