+++ packages/fighting-design/image-preview/src/props.ts
@@ -1,5 +1,5 @@
 import type { PropType, ExtractPropTypes } from 'vue'
-import type { ImagePreviewOnClose } from './interface'
+import type { HandleMouse } from '../../_interface'
 
 export const Props = {
   /**
@@ -47,10 +47,7 @@ export const Props = {
    */
   zIndex: {
     type: Number,
-    default: (): number => 999,
-    validator: (val: number): boolean => {
-      return val >= 0
-    }
+    default: (): null => null
   },
   /**
    * 是否展示图片预览下面的操作栏
@@ -70,7 +67,7 @@ export const Props = {
    * 关闭之后的回调
    */
   onClose: {
-    type: Function as PropType<ImagePreviewOnClose>,
+    type: Function as PropType<HandleMouse>,
     default: (): null => null
   }
 } as const