+++ packages/fighting-design/layout/src/interface.d.ts
@@ -1,3 +1,6 @@
 export type { LayoutProps } from './props'
 
-export type LayoutDirection = 'horizontal' | 'vertical' | ''
+/**
+ * 排列方向
+ */
+export type LayoutDirection = 'horizontal' | 'vertical'