+++ packages/fighting-design/link/src/interface.d.ts
@@ -12,4 +12,4 @@ export type LinkTarget = '_self' | '_blank' | '_parent' | '_top'
 /**
  * 不同 hover 状态
  */
-export type LinkHover = 'line' | 'bag' | ''
+export type LinkState = 'line' | 'bag'